Move hover service into core workbench
Fixes #100637
