+++ src/vs/workbench/services/hover/browser/hover.ts
