+++ src/vs/workbench/services/hover/browser/media/hover.css
